requirements.txt
Flask==3.0.0
mysql_connector_repackaged==0.3.1
pandas==2.1.4
PyMySQL==1.1.0
Requests==2.31.0
scikit_learn==1.3.2
xgboost==2.0.2
